Fix compiling error on OSX when using --enable-debug
In file included from globals.cpp:11:
In file included from ./addrman.h:12:
In file included from ./sync.h:10:
./deadlock-detection/threaddeadlock.h:34:22: error: no member named 'thread' in namespace 'std'; did you mean 'boost::thread'?
    static std::hash<std::thread::id> hasher;
                     ^~~~~~~~~~~
                     boost::thread

+++ src/unlimited.cpp
@@ -2169,7 +2169,7 @@ extern UniValue getstructuresizes(const UniValue &params, bool fHelp)
 #ifdef DEBUG_LOCKORDER
     {
         std::lock_guard<std::mutex> lock(lockdata.dd_mutex);
-        ret.pushKV("lockorders", lockdata.ordertracker.size());
+        ret.pushKV("lockorders", (uint64_t)lockdata.ordertracker.size());
     }
 #endif
     LOCK(cs_vNodes);
